+++ src/api/apiStartVideoCall.ts
@@ -3,7 +3,7 @@ import { fetchData, FETCH_METHODS, FETCH_SUCCESS } from './fetchData';
 
 export const apiStartVideoCall = async (
 	sessionId: number
-): Promise<{ videoCallUrl: string }> => {
+): Promise<{ moderatorVideoCallUrl: string }> => {
 	const url = config.endpoints.startVideoCall;
 	const videoCallData = JSON.stringify({
 		sessionId: sessionId