+++ src/api/index.ts
@@ -18,6 +18,7 @@ export * from './apiLogoutKeycloak';
 export * from './apiLogoutRocketchat';
 export * from './apiPostRegistration';
 export * from './apiPutGroupChat';
+export * from './apiPutConsultantData';
 export * from './apiPutEmail';
 export * from './apiPutSessionData';
 export * from './apiRegistrationNewConsultingTypes';